refactor: extract shared ViewHierarchyDumper for snapshot tests
The viewHierarchy/dumpHierarchy helper was duplicated across all 3
control-visibility test files with slightly different signatures.
Extract into a single ViewHierarchyDumper enum so all tests use
identical hosting/layout/dump behaviour.

Addresses Copilot review feedback on PR #223.
